requirements.txt
google-genai-adk==0.1.0
google-generativeai>=0.8.0
Pillow>=10.0.0
customtkinter>=5.2.0
